Work on Welcome Screen
Initial work on the welcome screen and required resource files for its completion. The string and color resources are modified to include the string values and global color scheme to be used respectively. Text attributes for the TextView components of the XML layout are changed through a function in the Activity class, and an Intent to move to next Activity upon thread completion is created