Modified SQLite Class for Testing Optimization
Deleted the class dedicated for the extension of the SQLite Open Helper superclass and attributes, constructors, and methods migrated to the DatabaseMethodHelper class which is then refactored to become the sole helper class for the database application. This is done in an effort to simplify the database unit test and not be able to use the Mockito class for unit testing.