Fix for User ID Error
Fixed ID of user data being of equal to 0 regardless of position of row in table. Turns out that the ID in the model class has an empty value and requires to be intentionally defined by converting the long return value of the query for adding and setting it as value for the variable. This fixes the error present in the tests