Finish Unit Test for UserInformationTableDAO
Written and successfully ran the unit tests for the data access object for the user_information_table model class. The tests checked whether the behavior of the DAO methods match that of the expected intentions of the developer, all of which returns as passed tests. Further cleanup of the code is done to optimize the code.